server/requirements.txt
openenv-core[core] @ git+https://github.com/meta-pytorch/OpenEnv.git@v0.2.3
fastmcp
torch>=2.0.0
gymnasium>=0.29.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
